requirements.txt
streamlit
firebase-admin
requests
python-dotenv
google-cloud-firestore